requirements.txt
PyQt5
matplotlib
pyyaml
